simple JSON using parse,  stringify, encodeURIComponent methods.
--in Server side App.js--
We are using the app.get() method to process the form.
request.query will hold the querystring parameters. The callback function expects a "strJSON" parameter in the querystring and assigns its value to strJSON:
const strJSON = request.query['strJSON'];
We use JSON.parse() to convert strJSON to an object and assign that object to objJSON.
We set the value of responseMsg based on the value of the "msg" key of objJSON and then we send responseMsg back to the client.
--In Client Side Ajax.js--
When the page loads, we create event listeners for clicks on the "hi" and "bye" buttons. The callback functions for both call sendRequest() and pass in the innerHTML of the button.
The sendRequest() function does the following:
Creates a simple JSON object with one property holding the passed in msg string.
Converts the JSON object to a string and encodes it for passing to the server using the encodeURIComponent() method.
Passes the string as a parameter of our Ajax request, which uses the GET method.
Outputs the responseText to the page.